resolving errors + switching some classes to interfaces
