+++ src/main/java/com/example/BookstoreBakersButBetter/collection/book.java
@@ -1,6 +1,8 @@
-package collection;
+package com.example.BookstoreBakersButBetter.collection;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import lombok.Builder;
 import lombok.Data;
 import lombok.Getter;
@@ -18,12 +20,13 @@ import org.springframework.data.mongodb.core.mapping.Document;
 
 
 public class book {
-    private static final java.lang.Object GenerationType = AUTO ;
+  //  private static final java.lang.Object GenerationType = AUTO;
     @Id
-    private String bookID;
     @GeneratedValue(strategy = GenerationType.AUTO)
+    private String bookID;
 
-    private string title;
+
+    private String title;
     private String publisher;
     private String author;
     private String publishDate;
