+++ src/main/java/com/example/BookstoreBakersButBetter/controller/RequestBody.java
@@ -1,0 +1,4 @@
+package com.example.BookstoreBakersButBetter.controller;
+
+public @interface RequestBody {
+}
