+++ src/main/java/com/example/BookstoreBakersButBetter/controller/bookController.java
@@ -1,6 +1,7 @@
-package controller;
+package com.example.BookstoreBakersButBetter.controller;
 
-import collection.book;
+import com.example.BookstoreBakersButBetter.collection.book;
+import com.example.BookstoreBakersButBetter.service.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,7 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/book")
 public class bookController {
     @Autowired
-    private bookService bookService;
+    private BookService bookService;
 
     @Override
     public String save(@RequestBody book book){
