+++ src/main/java/com/example/BookstoreBakersButBetter/repository/BookRepository.java
@@ -1,0 +1,9 @@
+package com.example.BookstoreBakersButBetter.repository;
+
+import com.example.BookstoreBakersButBetter.collection.book;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BookRepository extends MongoRepository<book,String> {
+}
