+++ src/main/java/com/example/BookstoreBakersButBetter/repository/ReviewCommentsRepository.java
@@ -1,0 +1,10 @@
+package com.example.BookstoreBakersButBetter.repository;
+
+import com.example.BookstoreBakersButBetter.collection.reviewComments;
+import com.example.BookstoreBakersButBetter.collection.reviews;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ReviewCommentsRepository extends MongoRepository<reviewComments,String> {
+}
