+++ src/main/java/com/example/BookstoreBakersButBetter/service/BookService.java
@@ -1,0 +1,7 @@
+package com.example.BookstoreBakersButBetter.service;
+
+import com.example.BookstoreBakersButBetter.collection.book;
+
+public interface BookService {
+    Object save(book book);
+}
