--- src/main/java/controller/Override.java
@@ -1,4 +1,0 @@
-package controller;
-
-public @interface Override {
-}
