--- src/main/java/controller/RequestBody.java
@@ -1,4 +1,0 @@
-package controller;
-
-public @interface RequestBody {
-}
