--- src/main/java/repository/bookRepository.java
@@ -1,19 +1,0 @@
-package repository;
-
-import collection.book;
-import org.springframework.data.domain.Example;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.data.repository.query.FluentQuery;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.function.Function;
-
-@Repository
-public class bookRepository extends MongoRepository <book,String> {
-
-}
